% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IowaHomesPredictionsShinySQL.r
\name{db_connection}
\alias{db_connection}
\title{Connecting to a database}
\usage{
db_connection()
}
\description{
This function will connect to the SQL database in order to query the information stored. The function will long and get the required data, after this it will create a dataframe with the chosen columns for the prediction.
}
