% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IowaHomesPredictionsShinySQL.r
\name{get_prediction}
\alias{get_prediction}
\title{Getting a prediction}
\usage{
get_prediction(input)
}
\description{
This function includes everything needed to make a prediction from a set of data. It uses the kknn method to get a model then after creating a data frame we use predict() with our model and dataframe to get our prediction. The function then rounds the output.
}
